Removed IGitt from manager base class
